python-selenium-automation/requirements.txt
selenium
pandas
openpyxl
python-dotenv
chromedriver-autoinstaller